refactor(cli): extract file I/O helpers into copybook-cli-io crate
