+++ Cargo.toml
@@ -18,6 +18,7 @@ members = [
     "copybook-codec",
     "copybook-codec-memory",
     "copybook-sequence-ring",
+    "copybook-cli-io",
     "copybook-cli",
     "copybook-gen",
     "copybook-governance",
@@ -68,6 +69,7 @@ copybook-dialect = { version = "=0.4.3", path = "copybook-dialect" }
 copybook-codec = { version = "=0.4.3", path = "copybook-codec" }
 copybook-codec-memory = { version = "=0.4.3", path = "copybook-codec-memory" }
 copybook-sequence-ring = { version = "=0.4.3", path = "copybook-sequence-ring" }
+copybook-cli-io = { version = "=0.4.3", path = "copybook-cli-io" }
 copybook-arrow = { version = "=0.4.3", path = "copybook-arrow" }
 copybook-governance = { version = "=0.4.3", path = "copybook-governance" }
 copybook-governance-grid = { version = "=0.4.3", path = "copybook-governance-grid" }
