+++ copybook-cli/Cargo.toml
@@ -39,6 +39,7 @@ chrono.workspace = true
 anyhow.workspace = true
 copybook-core = { workspace = true }
 copybook-codec = { workspace = true }
+copybook-cli-io = { workspace = true }
 copybook-governance = { workspace = true }
 copybook-arrow = { workspace = true, optional = true }
 tokio = { workspace = true, optional = true }
