+++ copybook-cli/src/commands/parse.rs
@@ -2,8 +2,8 @@
 //! Parse command implementation
 
 use crate::exit_codes::ExitCode;
-use crate::utils::{atomic_write, read_file_or_stdin};
 use crate::write_stdout_all;
+use copybook_cli_io::{atomic_write, read_file_or_stdin};
 use copybook_core::{ParseOptions, parse_copybook_with_options};
 use std::path::PathBuf;
 use tracing::info;
